GitHubCI: move 'publish' workflow inside the 'build' one
This way the 'publish' workflow is tested in every build, to make sure
it doesn't break (before this, we would only notice of any breakages
at the moment we create the tag). To prevent a release and a nuget
upload to happen at every commit, we just append an 'if' (see [1] for
the syntax) element to these steps.

[1] https://github.community/t/if-not-startswith-mutually-exclusive-steps/141841/2