.config/dotnet-tools: remove recursive dependency
This repo depended on the nuget package of itself, I guess
because we used to have some CI step that checked FSharpLint's
code against itself. We seem to not have that step anymore,
but if we want to revive it, we should do it with version
of FSharpLint that is compiled in the CI itself, to not have
a recursive dependency on older versions.