CI,build.fsx: publish prerelease nuget version for each commit
For this, the "alpha01" prerelease name is replaced by a dynamic one
that contains the date and the git hash.

We've had to move the Helpers before the Variables in order to use
them in this new part of the script.

As a side-node, these are nice FAKE docs about process handling here:
https://fake.build/reference/fake-core-process.html