build.fsx,RELEASE.md: improve preRelease version numbers
After we added[1] support for uploading prerelease nuget packages,
the versions chosen for these would increment the "Minor" part of
the last version released. Example: if last released version was
0.21.7, then the prerelease uploaded would start as something like
'0.22.0-date...'.

But this was not a good strategy because if the next version
released after 0.21.7 is 0.21.8, then the 0.22.x packages would
still be listed earlier than the 0.21.8 package in nuget, falsely
implying that they are "ahead" of 0.21.8, when actually they were
somewhere between 0.21.7 and 0.21.8.

Therefore we now make the increment happen at the Revision level
instead of at the Minor level. So the prereleases after releasing
a "0.21.8" for example, will start as '0.21.9-date...'.

And to be more in line with this change, we're gonna start the
convention of adding +2 to the Revision every time we tag a
release. This way, even numbers will give a "stable" aspect (as
they are not a prerelease anyway): 0.21.8, 0.21.10, 0.21.12...
And odd numbers will be prereleases: 0.21.9-date..., 0.21.11-date...

[1] 9618a8aa532b2dadc47ff5002d0adf678865c889