build.fsx,RELEASE.md: don't push prereleases for relNotes commits
The way we had setup the prerelease push was making unnecessary prerelease
packages for the same commit that ends up being mapped to a release. But if
we push the tag and the branch at the same time, and check the commit with
'git describe' then we can detect the situation and avert it.