docs,README: improve/elaborate main prj description [no ci]
By just describing as 'style checking tool', one could have thought
of FSharpLint as a formatting tool (sort of an alternative to fantomas)
but we cover way more than that. And to give a better idea of what
kind of issues can be detected, we put three different kinds of rules
here.