FsLint.Console: change RollForward to LatestMajor
It looks like the current roll foward behaviour was added
for fsprojects#519 to fix issues where only a newer version than the
tool was built with, is available.

Changing this to LatestMajor, however, would make it just use
the latest all the time (so for example, if you have both .NET6
and .NET8 installed, then it will use .NET8 all the time).

Fixes fsprojects#687