Makefile: don't skip 'docs' target
If there is a file or directory named the same as target
('docs' in this case) in current directory, Make treats that
target as a file target. Since the file (dir in this case)
already exists and has no dependencies (or its dependencies
are older), Make considered it "up to date" and skipped execution.

To avoid this, we now use a .PHONY target to tell Make that 'docs'
is not a real file.

Fixes https://github.com/fsprojects/FSharpLint/issues/768