Makefile: don't skip 'docs' target
If there is a file or directory named the same as target
('docs' in this case) in current directory, Make treats that
target as a file target. Since the file already exists and has
no dependencies (or its dependencies are older), Make considers
it "up to date" and skips execution.

Used a .PHONY target to tell Make that 'docs' is not a real
file.