Update Microsoft.Build dependencies
The versions its currently using are listed as suffering from https://github.com/advisories/GHSA-w3q9-fxm7-j8fq.

When I try to build the current code locally, it fails to build because the build has warnings-as-errors enabled, and NuGet package auditing flags up the CVE as an error.

So - maybe the dependencies need to be updated to the fixed versions?