Consolidate error/warning/suggestion terms into violation/failure
When using FSharpLint it's usual to refer to the reports that the rules
generate as "errors" or "warnings" or "suggestions", and the first two
are easily confused with compiler errors/warnings if not careful. Using
so many different terms is also confusing so with this commit we will
from now on call them "violations". Note, however, that some instances
of "error" have been replaced with "failures" instead of "violations" if
they refer to exceptions or other bugs/issues that FSharpLint might
generate when not working properly.