Rename FL75 rule
Being a rule that is related to naming, it was moved recently to the
"Naming" subcategory (and folder) inside Conventions; however, all other
rules in that folder contained rules that are finding elements depending
on the element type (e.g. NestedFunctionNames, PrivateValuesNames, etc),
which is not exactly what this rule is doing. So to better differentiate
this kind of rule, we change its suffix from "Names" to "Naming".